@@ -11,10 +11,11 @@ import PublicSharpIcon from '@mui/icons-material/PublicSharp';
 import SyncOutlinedIcon from '@mui/icons-material/SyncOutlined';
 import ThreeDRotationIcon from '@mui/icons-material/ThreeDRotation';
 import VideoCameraBackIcon from '@mui/icons-material/VideoCameraBackOutlined';
-import { WebSocketContext } from '../WebSocket/WebSocket';
-import RenderWindow from '../RenderWindow/RenderWindow';
-
-const msgpack = require('msgpack-lite');
+import { isEqual } from 'lodash';
+import {
+  makeThrottledMessageSender,
+  ViserWebSocketContext,
+} from '../WebSocket/ViserWebSocket';
 
 function CameraToggle() {
   const dispatch = useDispatch();
@@ -108,7 +109,7 @@ export default function ViewerWindow(props) {
   const labelRenderer = sceneTree.metadata.labelRenderer;
 
   const myRef = useRef(null);
-  const websocket = useContext(WebSocketContext).socket;
+  const viser_websocket = useContext(ViserWebSocketContext);
   const field_of_view = useSelector(
     (state) => state.renderingState.field_of_view,
   );
@@ -179,13 +180,28 @@ export default function ViewerWindow(props) {
     return () => {
       window.removeEventListener('resize', handleNewDimensions);
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   // start the three.js rendering loop
   // when the DOM is ready
   useEffect(() => {
+    document.getElementById("background-image").onload = function () {
+      if (scene) {
+        const oldBackground = scene.background;
+        const texture = new THREE.Texture(this);
+        texture.minFilter = THREE.LinearFilter;
+        texture.magFilter = THREE.NearestFilter;
+        texture.needsUpdate = true;
+        scene.background = texture;
+        if (oldBackground) {
+          oldBackground.dispose();
+        }
+      }
+    }
     myRef.current.append(renderer.domElement);
     render();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const render_height = useSelector(
@@ -199,17 +215,17 @@ export default function ViewerWindow(props) {
   let crop_h;
   const render_aspect = render_width / render_height;
   const viewport_aspect = viewport_width / viewport_height;
-  let render_viewport_apsect_ratio = null;
+  let render_viewport_aspect_ratio = null;
   if (render_aspect > viewport_aspect) {
     // render width is the limiting factor
     crop_w = viewport_width;
     crop_h = viewport_width / render_aspect;
-    render_viewport_apsect_ratio = viewport_aspect / render_aspect;
+    render_viewport_aspect_ratio = viewport_aspect / render_aspect;
   } else {
     // render height is the limiting factor
     crop_w = viewport_height * render_aspect;
     crop_h = viewport_height;
-    render_viewport_apsect_ratio = 1.0;
+    render_viewport_aspect_ratio = 1.0;
   }
 
   let display = null;
@@ -229,34 +245,41 @@ export default function ViewerWindow(props) {
   // such that the rendered video will match correctly
   if (camera_choice !== 'Main Camera') {
     const fl = 1.0 / Math.tan((field_of_view * Math.PI) / 360);
-    const fl_new = fl * render_viewport_apsect_ratio;
+    const fl_new = fl * render_viewport_aspect_ratio;
     const fov = Math.atan(1 / fl_new) / (Math.PI / 360);
     sceneTree.metadata.camera.fov = fov;
   } else {
     sceneTree.metadata.camera.fov = 50;
   }
 
+  let old_camera_matrix = null;
+  let is_moving = false;
+  const sendThrottledCameraMessage = makeThrottledMessageSender(
+    viser_websocket,
+    25,
+  );
   // update the camera information in the python server
   const sendCamera = () => {
-    if (websocket.readyState === WebSocket.OPEN) {
-      const cmd = 'write';
-      const path = 'renderingState/camera';
-      const data_packet = sceneTree.metadata.camera.toJSON();
-      data_packet.object.timestamp = +new Date();
-      if (camera_choice === 'Render Camera') {
-        data_packet.object.camera_type = camera_type;
+    if (isEqual(old_camera_matrix, sceneTree.metadata.camera.matrix.elements)) {
+      if (is_moving) {
+        is_moving = false;
       } else {
-        data_packet.object.camera_type = 'perspective';
+        return;
       }
-      data_packet.object.render_aspect = render_aspect;
-      const data = {
-        type: cmd,
-        path,
-        data: data_packet,
-      };
-      const message = msgpack.encode(data);
-      websocket.send(message);
+    } else {
+      is_moving = true;
     }
+    old_camera_matrix = sceneTree.metadata.camera.matrix.elements.slice();
+    sendThrottledCameraMessage({
+      type: 'CameraMessage',
+      aspect: sceneTree.metadata.camera.aspect,
+      render_aspect,
+      fov: sceneTree.metadata.camera.fov,
+      matrix: old_camera_matrix,
+      camera_type,
+      is_moving,
+      timestamp: +new Date(),
+    });
   };
 
   // keep sending the camera often
@@ -268,11 +291,41 @@ export default function ViewerWindow(props) {
     return () => {
       clearInterval(refreshIntervalId);
     };
-  }, [websocket, camera_choice, camera_type, render_aspect]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [viser_websocket, camera_choice, camera_type, render_aspect]);
+
+  const isWebsocketConnected = useSelector(
+    (state) => state.websocketState.isConnected,
+  );
+  useEffect(() => {
+    if (isWebsocketConnected) {
+      sendThrottledCameraMessage({
+        type: 'CameraMessage',
+        aspect: sceneTree.metadata.camera.aspect,
+        render_aspect,
+        fov: sceneTree.metadata.camera.fov,
+        matrix: sceneTree.metadata.camera.matrix.elements.slice(),
+        camera_type,
+        is_moving: false,
+        timestamp: +new Date(),
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isWebsocketConnected]);
 
   return (
     <>
-      <RenderWindow />
+      <div className="RenderWindow">
+        <div id="not-connected-overlay" hidden={isWebsocketConnected}>
+          <div id="not-connected-overlay-text">Renderer Disconnected</div>
+        </div>
+      </div>
+      <img
+        id="background-image"
+        alt="Render window"
+        z-index="1"
+        hidden
+      />
       <div className="canvas-container-main" ref={myRef}>
         <div className="ViewerWindow-camera-toggle">
           <CameraToggle />
