@@ -48,6 +48,7 @@ from nerfstudio.data.dataparsers.instant_ngp_dataparser import (
     InstantNGPDataParserConfig,
 )
 from nerfstudio.data.dataparsers.minimal_dataparser import MinimalDataParserConfig
+from nerfstudio.data.dataparsers.mipnerf_dataparser import MipNerf360DataParserConfig
 from nerfstudio.data.dataparsers.nerfosr_dataparser import NeRFOSRDataParserConfig
 from nerfstudio.data.dataparsers.multicam_dataparser import MulticamDataParserConfig
 from nerfstudio.data.dataparsers.nerfstudio_dataparser import NerfstudioDataParserConfig
@@ -122,6 +123,7 @@ AnnotatedDataParserUnion = tyro.conf.OmitSubcommandPrefixes[  # Omit prefixes of
             "blender-data": BlenderDataParserConfig(),
             "adop-data": AdopDataParserConfig(),
             "multicam-data": MulticamDataParserConfig(),
+            "mipnerf360-data": MipNerf360DataParserConfig(),
             "instant-ngp-data": InstantNGPDataParserConfig(),
             "nuscenes-data": NuScenesDataParserConfig(),
             "dnerf-data": DNeRFDataParserConfig(),
@@ -165,10 +167,11 @@ class DataManager(nn.Module):
     Usage:
     To get data, use the next_train and next_eval functions.
     This data manager's next_train and next_eval methods will return 2 things:
-        1. A Raybundle: This will contain the rays we are sampling, with latents and
-            conditionals attached (everything needed at inference)
-        2. A "batch" of auxiliary information: This will contain the mask, the ground truth
-            pixels, etc needed to actually train, score, etc the model
+
+    1. A Raybundle: This will contain the rays we are sampling, with latents and
+        conditionals attached (everything needed at inference)
+    2. A "batch" of auxiliary information: This will contain the mask, the ground truth
+        pixels, etc needed to actually train, score, etc the model
 
     Rationale:
     Because of this abstraction we've added, we can support more NeRF paradigms beyond the
@@ -194,6 +197,7 @@ class DataManager(nn.Module):
         eval_count (int): the step number of our eval iteration, needs to be incremented manually
         train_dataset (Dataset): the dataset for the train dataset
         eval_dataset (Dataset): the dataset for the eval dataset
+        includes_time (bool): whether the dataset includes time information
 
         Additional attributes specific to each subclass are defined in the setup_train and setup_eval
         functions.
@@ -204,6 +208,7 @@ class DataManager(nn.Module):
     eval_dataset: Optional[Dataset] = None
     train_sampler: Optional[DistributedSampler] = None
     eval_sampler: Optional[DistributedSampler] = None
+    includes_time: bool = False
 
     def __init__(self):
         """Constructor for the DataManager class.
@@ -422,6 +427,7 @@ class VanillaDataManager(DataManager):  # pylint: disable=abstract-method
         else:
             self.config.data = self.config.dataparser.data
         self.dataparser = self.dataparser_config.setup()
+        self.includes_time = self.dataparser.includes_time
         self.train_dataparser_outputs = self.dataparser.get_dataparser_outputs(split="train")
 
         self.train_dataset = self.create_train_dataset()
