@@ -17,19 +17,21 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Type
+from typing import Type, List
 
 import imageio
 import numpy as np
 import torch
 
 from nerfstudio.cameras.cameras import Cameras, CameraType
+from nerfstudio.data.dataparsers.adop_dataparser import TRAIN_INDICES, WEIGHTS
 from nerfstudio.data.dataparsers.base_dataparser import (
     DataParser,
     DataParserConfig,
     DataparserOutputs,
 )
 from nerfstudio.data.scene_box import SceneBox
+from nerfstudio.data.utils.colmap_parsing_utils import read_cameras_binary
 
 
 @dataclass
@@ -40,14 +42,15 @@ class MipNerf360DataParserConfig(DataParserConfig):
     """target class to instantiate"""
     data: Path = Path("data/mipnerf_360/garden")
     """Directory specifying location of data."""
-    downscale_factor: int = 1
+    scales: List[int] = field(default_factory=lambda: [1, 2, 4, 8])
     """How much to downscale images."""
     val_skip: int = 8
     """1/val_skip images to use for validation."""
     auto_scale: bool = True
     """Scale based on pose bounds."""
-    aabb_scale: float = 4
+    aabb_scale: float = 1
     """Scene scale."""
+    train_split: float = 7 / 8
 
 
 @dataclass
@@ -78,86 +81,147 @@ class Mipnerf360(DataParser):
         return poses_orig
 
     def _generate_dataparser_outputs(self, split="train"):
-        image_dir = "images"
-        if self.config.downscale_factor > 1:
-            image_dir += f"_{self.config.downscale_factor}"
-        image_dir = self.config.data / image_dir
-        if not image_dir.exists():
-            raise ValueError(f"Image directory {image_dir} doesn't exist")
-
-        valid_formats = [".jpg", ".png"]
+        fx = []
+        fy = []
+        cx = []
+        cy = []
+        c2ws = []
+        width = []
+        height = []
+        weights = []
+        img_scales = []
         image_filenames = []
-        for f in image_dir.iterdir():
-            ext = f.suffix
-            if ext.lower() not in valid_formats:
-                continue
-            image_filenames.append(image_dir / f)
-        image_filenames = sorted(image_filenames)
-        num_images = len(image_filenames)
 
-        poses_data = np.load(self.config.data / "poses_bounds.npy")
-        poses = poses_data[:, :-2].reshape([-1, 3, 5]).astype(np.float32)
-        bounds = poses_data[:, -2:].transpose([1, 0])
+        camera_params = read_cameras_binary(self.config.data / 'sparse/0/cameras.bin')
+        assert camera_params[1].model == 'PINHOLE'
+        camera_fx, camera_fy, camera_cx, camera_cy = camera_params[1].params
 
-        if num_images != poses.shape[0]:
-            raise RuntimeError(f"Different number of images ({num_images}), and poses ({poses.shape[0]})")
+        for scale in self.config.scales:
+            image_dir = "images"
+            if scale > 1:
+                image_dir += f"_{scale}"
+            image_dir = self.config.data / image_dir
+            if not image_dir.exists():
+                raise ValueError(f"Image directory {image_dir} doesn't exist")
 
-        idx_test = np.arange(num_images)[:: self.config.val_skip]
-        idx_train = np.array([i for i in np.arange(num_images) if i not in idx_test])
-        idx = idx_train if split == "train" else idx_test
+            valid_formats = [".jpg", ".png"]
+            num_images = 0
+            for f in sorted(image_dir.iterdir()):
+                ext = f.suffix
+                if ext.lower() not in valid_formats:
+                    continue
+                image_filenames.append(f)
+                num_images += 1
 
-        image_filenames = np.array(image_filenames)[idx]
-        poses = poses[idx]
+            poses_data = np.load(self.config.data / "poses_bounds.npy")
+            poses = poses_data[:, :-2].reshape([-1, 3, 5]).astype(np.float32)
+            bounds = poses_data[:, -2:].transpose([1, 0])
 
-        img_0 = imageio.imread(image_filenames[0])
-        image_height, image_width = img_0.shape[:2]
+            if num_images != poses.shape[0]:
+                raise RuntimeError(f"Different number of images ({num_images}), and poses ({poses.shape[0]})")
 
-        poses[:, :2, 4] = np.array([image_height, image_width])
-        poses[:, 2, 4] = poses[:, 2, 4]
+            img_0 = imageio.imread(image_filenames[-1])
+            image_height, image_width = img_0.shape[:2]
 
-        # Reorder pose to match our convention
-        poses = np.concatenate([poses[:, :, 1:2], -poses[:, :, 0:1], poses[:, :, 2:]], axis=-1)
+            width.append(torch.full((num_images, 1), image_width, dtype=torch.long))
+            height.append(torch.full((num_images, 1), image_height, dtype=torch.long))
+            fx.append(torch.full((num_images, 1), camera_fx / scale))
+            fy.append(torch.full((num_images, 1), camera_fy / scale))
+            cx.append(torch.full((num_images, 1), camera_cx / scale))
+            cy.append(torch.full((num_images, 1), camera_cy / scale))
+            weights.append(torch.full((num_images, 1), scale ** 2))
+            img_scales.append(torch.full((num_images, 1), scale, dtype=torch.long))
 
-        # Center poses and rotate. (Compute up from average of all poses)
-        poses = self.normalize_orientation(poses)
+            # Reorder pose to match our convention
+            poses = np.concatenate([poses[:, :, 1:2], -poses[:, :, 0:1], poses[:, :, 2:]], axis=-1)
 
-        # Scale factor used in mipnerf
-        if self.config.auto_scale:
-            scale_factor = 1 / (np.min(bounds) * 0.75)
-            poses[:, :3, 3] *= scale_factor
-            bounds *= scale_factor
+            # Center poses and rotate. (Compute up from average of all poses)
+            poses = self.normalize_orientation(poses)
 
-        # Center poses
-        poses[:, :3, 3] = poses[:, :3, 3] - np.mean(poses[:, :3, :], axis=0)[:, 3]
+            # Scale factor used in mipnerf
+            if self.config.auto_scale:
+                scale_factor = 1 / (np.min(bounds) * 0.75)
+                poses[:, :3, 3] *= scale_factor
+                bounds *= scale_factor
 
-        focal_length = poses[0, -1, -1]
+            # Center poses
+            poses[:, :3, 3] = poses[:, :3, 3] - np.mean(poses[:, :3, :], axis=0)[:, 3]
+            c2ws.append(torch.from_numpy(poses[:, :3, :4]))
 
-        cx = image_width / 2.0
-        cy = image_height / 2.0
-        camera_to_world = torch.from_numpy(poses[:, :3, :4])  # camera to world transform
-        num_cameras = len(image_filenames)
-        num_intrinsics_params = 3
-        intrinsics = torch.ones((num_cameras, num_intrinsics_params), dtype=torch.float32)
-        intrinsics *= torch.tensor([cx, cy, focal_length])
+        c2ws = torch.cat(c2ws)
+        min_bounds = c2ws[:, :, 3].min(dim=0)[0]
+        max_bounds = c2ws[:, :, 3].max(dim=0)[0]
 
-        aabb = torch.tensor([[-4, -4, -4], [4, 4, 4]], dtype=torch.float32) * self.config.aabb_scale
-        scene_box = SceneBox(aabb=aabb)
+        origin = (max_bounds + min_bounds) * 0.5
+        print('Calculated origin: {} {} {}'.format(origin, min_bounds, max_bounds))
+
+        pose_scale_factor = ((max_bounds - min_bounds) * 0.5).norm().item()
+        print('Calculated pose scale factor: {}'.format(pose_scale_factor))
+
+        for c2w in c2ws:
+            c2w[:, 3] = (c2w[:, 3] - origin) / pose_scale_factor
+            assert torch.logical_and(c2w >= -1, c2w <= 1).all(), c2w
+
+        # in x,y,z order
+        scene_box = SceneBox(aabb=((torch.stack([min_bounds, max_bounds]) - origin) / pose_scale_factor).float())
+
+        train_indices = set()
+        base_train_indices = np.linspace(0, num_images, int(num_images * self.config.train_split),
+                                         endpoint=False, dtype=np.int32)
+        for i in range(len(self.config.scales)):
+            train_indices.update(base_train_indices + num_images * i)
+
+        img_scales = torch.cat(img_scales)
+
+        if split.casefold() == 'train':
+            mask_filenames = []
+            for i in range(len(image_filenames)):
+                if i in train_indices:
+                    mask_filenames.append(self.config.data / f'image_full-{img_scales[i].item()}.png')
+                else:
+                    mask_filenames.append(self.config.data / f'image_left-{img_scales[i].item()}.png')
+
+            indices = torch.arange(len(image_filenames), dtype=torch.long)
+        else:
+            val_indices = []
+            mask_filenames = []
+            train_indices = set(train_indices)
+            for i in range(len(image_filenames)):
+                if i not in train_indices:
+                    val_indices.append(i)
+                    mask_filenames.append(self.config.data / f'image_right-{img_scales[i].item()}.png')
+
+            indices = torch.LongTensor(val_indices)
 
         cameras = Cameras(
-            fx=focal_length,
-            fy=focal_length,
-            cx=cx,
-            cy=cy,
-            camera_to_worlds=camera_to_world,
+            camera_to_worlds=c2ws[indices].float(),
+            fx=torch.cat(fx)[indices],
+            fy=torch.cat(fy)[indices],
+            cx=torch.cat(cx)[indices],
+            cy=torch.cat(cy)[indices],
+            width=torch.cat(width)[indices],
+            height=torch.cat(height)[indices],
             camera_type=CameraType.PERSPECTIVE,
         )
 
-        cameras.rescale_output_resolution(scaling_factor=1.0 / self.config.downscale_factor)
+        print('Num images in split {}: {}'.format(split, len(indices)))
+
+        embedding_indices = torch.arange(num_images).unsqueeze(0).repeat(len(self.config.scales), 1).view(-1)
+
+        metadata = {
+            TRAIN_INDICES: embedding_indices[indices],
+            WEIGHTS: torch.cat(weights)[indices],
+            "pose_scale_factor": pose_scale_factor,
+            "cameras": cameras,
+        }
 
         dataparser_outputs = DataparserOutputs(
-            image_filenames=image_filenames,
+            image_filenames=[image_filenames[i] for i in indices],
             cameras=cameras,
             scene_box=scene_box,
+            dataparser_scale=1,
+            mask_filenames=mask_filenames,
+            metadata=metadata
         )
 
         return dataparser_outputs
