@@ -42,7 +42,7 @@ from nerfstudio.data.scene_box import SceneBox
 CONSOLE = Console(width=120)
 
 
-def find_files(directory: str, exts: List[str]):
+def _find_files(directory: str, exts: List[str]):
     """Find all files in a directory that have a certain file extension.
 
     Args:
@@ -94,8 +94,8 @@ def get_camera_params(
     split_dir = f"{scene_dir}/{split}"
 
     # camera parameters files
-    intrinsics_files = find_files(f"{split_dir}/intrinsics", exts=["*.txt"])
-    pose_files = find_files(f"{split_dir}/pose", exts=["*.txt"])
+    intrinsics_files = _find_files(f"{split_dir}/intrinsics", exts=["*.txt"])
+    pose_files = _find_files(f"{split_dir}/pose", exts=["*.txt"])
 
     num_cams = len(pose_files)
 
@@ -219,12 +219,12 @@ class NeRFOSR(DataParser):
         )
 
         # --- images ---
-        image_filenames = find_files(f"{split_dir}/rgb", exts=["*.png", "*.jpg", "*.JPG", "*.PNG"])
+        image_filenames = _find_files(f"{split_dir}/rgb", exts=["*.png", "*.jpg", "*.JPG", "*.PNG"])
 
         # --- masks ---
         mask_filenames = []
         if self.config.use_masks:
-            mask_filenames = find_files(f"{split_dir}/mask", exts=["*.png", "*.jpg", "*.JPG", "*.PNG"])
+            mask_filenames = _find_files(f"{split_dir}/mask", exts=["*.png", "*.jpg", "*.JPG", "*.PNG"])
 
         dataparser_outputs = DataparserOutputs(
             image_filenames=image_filenames,
