@@ -18,7 +18,8 @@
 
 from abc import abstractmethod
 from dataclasses import dataclass, field
-from typing import Optional, Type, List
+
+from typing import Optional, Tuple, Type, List
 
 import numpy as np
 from torch.optim import Optimizer, lr_scheduler
@@ -55,6 +56,34 @@ class Scheduler:
         Returns:
             The scheduler object.
         """
+
+
+@dataclass
+class MultiStepSchedulerConfig(SchedulerConfig):
+    """Config for multi step scheduler where lr decays by gamma every milestone"""
+
+    _target: Type = field(default_factory=lambda: MultiStepScheduler)
+    """target class to instantiate"""
+    max_steps: int = 1000000
+    """The maximum number of steps."""
+    gamma: float = 0.33
+    """The learning rate decay factor."""
+    milestones: Tuple[int, ...] = (500000, 750000, 900000)
+    """The milestone steps at which to decay the learning rate."""
+
+
+class MultiStepScheduler(Scheduler):
+    """Multi step scheduler where lr decays by gamma every milestone"""
+
+    config: MultiStepSchedulerConfig
+
+    def get_scheduler(self, optimizer: Optimizer, lr_init: float) -> lr_scheduler._LRScheduler:
+        scheduler = lr_scheduler.MultiStepLR(
+            optimizer=optimizer,
+            milestones=self.config.milestones,
+            gamma=self.config.gamma,
+        )
+        return scheduler
 
 
 @dataclass
@@ -112,7 +141,7 @@ class ExponentialDecayScheduler(Scheduler):
 
 @dataclass
 class CosineDecaySchedulerConfig(SchedulerConfig):
-    """Basic scheduler config with self-defined exponential decay schedule"""
+    """Config for cosine decay schedule"""
 
     _target: Type = field(default_factory=lambda: CosineDecayScheduler)
     """target class to instantiate"""
@@ -125,7 +154,7 @@ class CosineDecaySchedulerConfig(SchedulerConfig):
 
 
 class CosineDecayScheduler(Scheduler):
-    """Starts with a flat lr schedule until it reaches N epochs then applies a given scheduler"""
+    """Cosine decay scheduler with linear warmup"""
 
     config: CosineDecaySchedulerConfig
 
