@@ -21,7 +21,6 @@ from typing import Dict, Optional, Tuple
 
 import numpy as np
 import torch
-from nerfacc import ContractionType, contract
 from torch.nn.parameter import Parameter
 from torchtyping import TensorType
 
@@ -30,6 +29,10 @@ from nerfstudio.data.scene_box import SceneBox
 from nerfstudio.field_components.activations import trunc_exp
 from nerfstudio.field_components.embedding import Embedding
 from nerfstudio.field_components.field_heads import FieldHeadNames
+from nerfstudio.field_components.spatial_distortions import (
+    SceneContraction,
+    SpatialDistortion,
+)
 from nerfstudio.fields.base_field import Field, shift_directions_for_tcnn
 
 try:
@@ -69,17 +72,17 @@ class TCNNInstantNGPField(Field):
         use_appearance_embedding: Optional[bool] = False,
         num_images: Optional[int] = None,
         appearance_embedding_dim: int = 32,
-        contraction_type: ContractionType = ContractionType.UN_BOUNDED_SPHERE,
         num_levels: int = 16,
         log2_hashmap_size: int = 19,
         max_res: int = 2048,
-        features_per_level: int = 8
+        features_per_level: int = 4,
+        spatial_distortion: Optional[SpatialDistortion] = SceneContraction(),
     ) -> None:
         super().__init__()
 
         self.aabb = Parameter(aabb, requires_grad=False)
         self.geo_feat_dim = geo_feat_dim
-        self.contraction_type = contraction_type
+        self.spatial_distortion = spatial_distortion
 
         self.use_appearance_embedding = use_appearance_embedding
         if use_appearance_embedding:
@@ -89,7 +92,8 @@ class TCNNInstantNGPField(Field):
 
         # TODO: set this properly based on the aabb
         base_res: int = 16
-        per_level_scale = np.exp((np.log(max_res) - np.log(base_res)) / (num_levels - 1))
+        features_per_level: int = 2
+        growth_factor = np.exp((np.log(max_res) - np.log(base_res)) / (num_levels - 1))
 
         self.direction_encoding = tcnn.Encoding(
             n_input_dims=3,
@@ -108,7 +112,7 @@ class TCNNInstantNGPField(Field):
                 "n_features_per_level": features_per_level,
                 "log2_hashmap_size": log2_hashmap_size,
                 "base_resolution": base_res,
-                "per_level_scale": per_level_scale,
+                "per_level_scale": growth_factor,
             },
             network_config={
                 "otype": "FullyFusedMLP",
@@ -135,9 +139,16 @@ class TCNNInstantNGPField(Field):
         )
 
     def get_density(self, ray_samples: RaySamples) -> Tuple[TensorType, TensorType]:
-        positions = ray_samples.frustums.get_positions()
+        if self.spatial_distortion is not None:
+            positions = ray_samples.frustums.get_positions()
+            positions = self.spatial_distortion(positions)
+            positions = (positions + 2.0) / 4.0
+        else:
+            positions = SceneBox.get_normalized_positions(ray_samples.frustums.get_positions(), self.aabb)
+        # Make sure the tcnn gets inputs between 0 and 1.
+        selector = ((positions > 0.0) & (positions < 1.0)).all(dim=-1)
+        positions = positions * selector[..., None]
         positions_flat = positions.view(-1, 3)
-        positions_flat = contract(x=positions_flat, roi=self.aabb, type=self.contraction_type)
 
         h = self.mlp_base(positions_flat).view(*ray_samples.frustums.shape, -1)
         density_before_activation, base_mlp_out = torch.split(h, [1, self.geo_feat_dim], dim=-1)
@@ -151,15 +162,12 @@ class TCNNInstantNGPField(Field):
     def get_outputs(
         self, ray_samples: RaySamples, density_embedding: Optional[TensorType] = None
     ) -> Dict[FieldHeadNames, TensorType]:
+        assert density_embedding is not None
         directions = shift_directions_for_tcnn(ray_samples.frustums.directions)
         directions_flat = directions.view(-1, 3)
-
         d = self.direction_encoding(directions_flat)
-        if density_embedding is None:
-            positions = SceneBox.get_normalized_positions(ray_samples.frustums.get_positions(), self.aabb)
-            h = torch.cat([d, positions.view(-1, 3)], dim=-1)
-        else:
-            h = torch.cat([d, density_embedding.view(-1, self.geo_feat_dim)], dim=-1)
+
+        h = torch.cat([d, density_embedding.view(-1, self.geo_feat_dim)], dim=-1)
 
         if self.use_appearance_embedding:
             if ray_samples.camera_indices is None:
