@@ -19,9 +19,8 @@ Collection of sampling strategies
 from abc import abstractmethod
 from typing import Callable, List, Optional, Tuple, Union
 
-import nerfacc
 import torch
-from nerfacc import OccupancyGrid
+from nerfacc import OccGridEstimator
 from torch import nn
 from torchtyping import TensorType
 
@@ -382,24 +381,27 @@ class VolumetricSampler(Sampler):
 
     def __init__(
         self,
-        occupancy_grid: Optional[OccupancyGrid] = None,
-        density_fn: Optional[Callable[[TensorType[..., 3], Optional[TensorType[..., 1]]], TensorType[..., 1]]] = None,
-        scene_aabb: Optional[TensorType[2, 3]] = None,
+# <<<<<<< HEAD
+#         occupancy_grid: Optional[OccupancyGrid] = None,
+#         density_fn: Optional[Callable[[TensorType[..., 3], Optional[TensorType[..., 1]]], TensorType[..., 1]]] = None,
+#         scene_aabb: Optional[TensorType[2, 3]] = None,
+# =======
+        occupancy_grid: OccGridEstimator,
+        density_fn: Optional[Callable[[TensorType[..., 3]], TensorType[..., 1]]] = None,
     ) -> None:
 
         super().__init__()
-        self.scene_aabb = scene_aabb
+        assert occupancy_grid is not None
         self.density_fn = density_fn
         self.occupancy_grid = occupancy_grid
-        if self.scene_aabb is not None:
-            self.scene_aabb = self.scene_aabb.to("cuda").flatten()
 
-    def get_sigma_fn(self, origins, directions, pixel_area) -> Optional[Callable]:
+    def get_sigma_fn(self, origins, directions, pixel_area, times=None) -> Optional[Callable]:
         """Returns a function that returns the density of a point.
 
         Args:
             origins: Origins of rays
             directions: Directions of rays
+            times: Times at which rays are sampled
         Returns:
             Function that returns the density of a point or None if a density function is not provided.
         """
@@ -412,10 +414,11 @@ class VolumetricSampler(Sampler):
         def sigma_fn(t_starts, t_ends, ray_indices):
             t_origins = origins[ray_indices]
             t_dirs = directions[ray_indices]
-            positions = t_origins + t_dirs * (t_starts + t_ends) / 2.0
-            de = density_fn(positions, None, origins=t_origins, directions=t_dirs, starts=t_starts, ends=t_ends,
-                            pixel_area=pixel_area[ray_indices])
-            return de
+
+            positions = t_origins + t_dirs * (t_starts + t_ends)[:, None] / 2.0
+            return density_fn(positions, times[ray_indices] if times is not None else None, origins=t_origins,
+                              directions=t_dirs, starts=t_starts[:, None], ends=t_ends[:, None],
+                              pixel_area=pixel_area[ray_indices]).squeeze(-1)
 
         return sigma_fn
 
@@ -431,8 +434,8 @@ class VolumetricSampler(Sampler):
         render_step_size: float,
         near_plane: float = 0.0,
         far_plane: Optional[float] = None,
+        alpha_thre: float = 0.01,
         cone_angle: float = 0.0,
-        alpha_thre: float = 1e-2,
     ) -> Tuple[RaySamples, TensorType["total_samples",]]:
         """Generate ray samples in a bounding box.
 
@@ -441,6 +444,7 @@ class VolumetricSampler(Sampler):
             render_step_size: Minimum step size to use for rendering
             near_plane: Near plane for raymarching
             far_plane: Far plane for raymarching
+            alpha_thre: Opacity threshold skipping samples.
             cone_angle: Cone angle for raymarching, set to 0 for uniform marching.
 
         Returns:
@@ -451,6 +455,7 @@ class VolumetricSampler(Sampler):
 
         rays_o = ray_bundle.origins.contiguous()
         rays_d = ray_bundle.directions.contiguous()
+        times = ray_bundle.times
 
         if ray_bundle.nears is not None and ray_bundle.fars is not None:
             t_min = ray_bundle.nears.contiguous().reshape(-1)
@@ -460,19 +465,19 @@ class VolumetricSampler(Sampler):
             t_min = None
             t_max = None
 
+        if far_plane is None:
+            far_plane = 1e10
+
         if ray_bundle.camera_indices is not None:
             camera_indices = ray_bundle.camera_indices.contiguous()
         else:
             camera_indices = None
-
-        ray_indices, starts, ends = nerfacc.ray_marching(
+        ray_indices, starts, ends = self.occupancy_grid.sampling(
             rays_o=rays_o,
             rays_d=rays_d,
             t_min=t_min,
             t_max=t_max,
-            scene_aabb=self.scene_aabb,
-            grid=self.occupancy_grid,
-            sigma_fn=self.get_sigma_fn(rays_o, rays_d, ray_bundle.pixel_area.contiguous()),
+            sigma_fn=self.get_sigma_fn(rays_o, rays_d, ray_bundle.pixel_area.contiguous(), times),
             render_step_size=render_step_size,
             near_plane=near_plane,
             far_plane=far_plane,
@@ -486,8 +491,8 @@ class VolumetricSampler(Sampler):
             # create a single fake sample and update packed_info accordingly
             # this says the last ray in packed_info has 1 sample, which starts and ends at 1
             ray_indices = torch.zeros((1,), dtype=torch.long, device=rays_o.device)
-            starts = torch.ones((1, 1), dtype=starts.dtype, device=rays_o.device)
-            ends = torch.ones((1, 1), dtype=ends.dtype, device=rays_o.device)
+            starts = torch.ones((1,), dtype=starts.dtype, device=rays_o.device)
+            ends = torch.ones((1,), dtype=ends.dtype, device=rays_o.device)
 
         origins = rays_o[ray_indices]
         dirs = rays_d[ray_indices]
@@ -498,8 +503,8 @@ class VolumetricSampler(Sampler):
             frustums=Frustums(
                 origins=origins,
                 directions=dirs,
-                starts=starts,
-                ends=ends,
+                starts=starts[..., None],
+                ends=ends[..., None],
                 pixel_area=ray_bundle[ray_indices].pixel_area,
             ),
             camera_indices=camera_indices,
