@@ -79,8 +79,11 @@ class ProcessNuScenesMasks:
                 # Instead use BoxVisibility.NONE and make sure to rasterize box faces correctly
 
                 mask = np.ones((900, 1600), dtype=np.uint8)
-                for box in boxes:
+                # If is backcam, mask the truck of the ego vehicle
+                if camera == "CAM_BACK":
+                    mask[-100:] = 0
 
+                for box in boxes:
                     # Dont mask out static objects (static in all frames)
                     instance_token = nusc.get("sample_annotation", box.token)["instance_token"]
                     if not instances_is_dynamic[instance_token]:
