@@ -21,6 +21,7 @@ BLACKLIST = [
     "instant-ngp",
     "instant-ngp-bounded",
     "nerfacto",
+    "nerfacto-big",
     "volinga",
     "phototourism",
     "depth-nerfacto",
@@ -29,6 +30,7 @@ BLACKLIST = [
     "neus",
     "kplanes",
     "kplanes-dynamic",
+    "neus-facto",
 ]
 
 
@@ -51,11 +53,7 @@ def set_reduced_config(config: TrainerConfig):
     if hasattr(config.pipeline.model, "num_importance_samples"):
         config.pipeline.model.num_importance_samples = 4
     # remove viewer
-    config.viewer.enable = False
-
-    # model specific config settings
-    if config.method_name == "instant-ngp":
-        config.pipeline.model.field_implementation = "torch"
+    config.viewer.quit_on_train_completion = True
 
     return config
 
